requirements.txt
requests
moviepy
pillow
pyttsx3